+++ lib/site.comment.like.js
